test: #11 configure test for D1
- miniflareを導入
- D1用のrepoテストケース追加
